refactor: Separate validation from resource processors
