+++ galasa-parent/dev.galasa.framework.api.resources/src/main/java/dev/galasa/framework/api/resources/processors/GalasaPropertyProcessor.java
@@ -8,12 +8,10 @@ package dev.galasa.framework.api.resources.processors;
 import static dev.galasa.framework.api.common.ServletErrorMessage.*;
 import static dev.galasa.framework.api.common.resources.ResourceAction.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.servlet.http.HttpServletResponse;
 
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
 import dev.galasa.framework.api.beans.GalasaProperty;
@@ -23,13 +21,12 @@ import dev.galasa.framework.api.common.resources.CPSFacade;
 import dev.galasa.framework.api.common.resources.CPSNamespace;
 import dev.galasa.framework.api.common.resources.CPSProperty;
 import dev.galasa.framework.api.common.resources.ResourceAction;
-import dev.galasa.framework.api.common.resources.ResourceNameValidator;
+import dev.galasa.framework.api.resources.validators.GalasaPropertyValidator;
 import dev.galasa.framework.spi.ConfigurationPropertyStoreException;
 
 public class GalasaPropertyProcessor extends AbstractGalasaResourceProcessor implements IGalasaResourceProcessor {
 
     private CPSFacade cps;
-    static final ResourceNameValidator nameValidator = new ResourceNameValidator();
 
     public GalasaPropertyProcessor(CPSFacade cps) {
         this.cps = cps;
@@ -75,61 +72,7 @@ public class GalasaPropertyProcessor extends AbstractGalasaResourceProcessor imp
     }
 
     private List<String> checkGalasaPropertyJsonStructure(JsonObject propertyJson, ResourceAction action) throws InternalServletException {
-        checkResourceHasRequiredFields(propertyJson, GalasaProperty.DEFAULTAPIVERSION, action);
-
-        List<String> validationErrors = new ArrayList<String>();
-        validatePropertyMetadata(propertyJson, validationErrors);
-
-        // Delete operations shouldn't require a 'data' section, just the metadata to identify
-        // the property to delete
-        if (action != DELETE) {
-            validatePropertyData(propertyJson, validationErrors);
-        }
-        return validationErrors;
-    }
-
-    private void validatePropertyMetadata(JsonObject propertyJson, List<String> validationErrors) {
-        //Check metadata is not null and contains name and namespace fields in the correct format
-        JsonObject metadata = propertyJson.get("metadata").getAsJsonObject();
-        if (metadata.has("name") && metadata.has("namespace")) {
-            JsonElement name = metadata.get("name");
-            JsonElement namespace = metadata.get("namespace"); 
-
-            // Use the ResourceNameValidator to check that the name is correctly formatted and not null
-            try {
-                nameValidator.assertPropertyNameCharPatternIsValid(name.getAsString());
-            } catch (InternalServletException e) {
-                // All ResourceNameValidator error should be added to the list of reasons why the property action has failed
-                validationErrors.add(e.getMessage());
-            }
-
-            // Use the ResourceNameValidator to check that the namespace is correctly formatted and not null
-            try {
-                nameValidator.assertNamespaceCharPatternIsValid(namespace.getAsString());
-            } catch (InternalServletException e) {
-                validationErrors.add(e.getMessage());   
-            }
-        } else {
-            String message = "The 'metadata' field cannot be empty. The fields 'name' and 'namespace' are mandatory for the type GalasaProperty.";
-            ServletError error = new ServletError(GAL5024_INVALID_GALASAPROPERTY, message);
-            validationErrors.add(new InternalServletException(error, HttpServletResponse.SC_BAD_REQUEST).getMessage());
-        }
-    }
-
-    private void validatePropertyData(JsonObject propertyJson, List<String> validationErrors) {
-        //Check that data is not null and contains the value field
-        JsonObject data = propertyJson.get("data").getAsJsonObject();
-        if (data.size() > 0 && data.has("value")) {
-            String value = data.get("value").getAsString();
-            if (value == null || value.isBlank()) {
-                String message = "The 'value' field cannot be empty. The field 'value' is mandatory for the type GalasaProperty.";
-                ServletError error = new ServletError(GAL5024_INVALID_GALASAPROPERTY, message);
-                validationErrors.add(new InternalServletException(error, HttpServletResponse.SC_BAD_REQUEST).getMessage());
-            }
-        } else {
-            String message = "The 'data' field cannot be empty. The field 'value' is mandatory for the type GalasaProperty.";
-            ServletError error = new ServletError(GAL5024_INVALID_GALASAPROPERTY, message);
-            validationErrors.add(new InternalServletException(error, HttpServletResponse.SC_BAD_REQUEST).getMessage());
-        }
+        GalasaPropertyValidator validator = new GalasaPropertyValidator(action);
+        return checkGalasaResourceJsonStructure(validator, propertyJson);
     }
 }
